[NixIO] Test skipping: {setUp,tearDown}Class methods
Class methods are not skipped when module is missing (by the unittest
decorator), so we need the checks to avoid failing when there is no NIX
module.

+++ neo/test/iotest/test_nixio.py
@@ -903,7 +903,8 @@ class NixIOReadTest(NixIOTest):
 
     @classmethod
     def setUpClass(cls):
-        cls.nixfile = cls.create_full_nix_file(cls.filename)
+        if HAVE_NIX:
+            cls.nixfile = cls.create_full_nix_file(cls.filename)
 
     def setUp(self):
         self.io = NixIO(self.filename, "ro")
@@ -912,8 +913,9 @@ class NixIOReadTest(NixIOTest):
 
     @classmethod
     def tearDownClass(cls):
-        cls.nixfile.close()
-        os.remove(cls.filename)
+        if HAVE_NIX:
+            cls.nixfile.close()
+            os.remove(cls.filename)
 
     def tearDown(self):
         self.io.close()
@@ -1109,7 +1111,8 @@ class NixIOPartialWriteTest(NixIOTest):
 
     @classmethod
     def setUpClass(cls):
-        cls.nixfile = cls.create_full_nix_file(cls.filename)
+        if HAVE_NIX:
+            cls.nixfile = cls.create_full_nix_file(cls.filename)
 
     def setUp(self):
         self.io = NixIO(self.filename, "rw")
@@ -1119,8 +1122,9 @@ class NixIOPartialWriteTest(NixIOTest):
 
     @classmethod
     def tearDownClass(cls):
-        cls.nixfile.close()
-        os.remove(cls.filename)
+        if HAVE_NIX:
+            cls.nixfile.close()
+            os.remove(cls.filename)
 
     def tearDown(self):
         self.restore_methods()
